refactor: 수리업체 프로필 UI 컴포넌트화 및 스크롤 소식, 후기 탭 스크롤 영역 설정
