+++ src/mocks/repair/vendors.js
@@ -44,6 +44,12 @@ export const MOCK_VENDORS = [
         title: '임시 휴무 (05/08)',
         body: '어버이날 맞이하여 고향에 내려갑니다 🙂 잠시 휴무입니다. 모두 즐거운 하루 되세요!',
       },
+      {
+        id: 'n3',
+        date: '2023.03.03',
+        title: '임시 휴무 (05/08)',
+        body: '어버이날 맞이하여 고향에 내려갑니다 🙂 잠시 휴무입니다. 모두 즐거운 하루 되세요!',
+      },
     ],
     reviews: [
       {
